fix(web): address PR review feedback - refactor rendering and fix data consistency
## Summary
Addressed critical and medium priority feedback from PR reviewers.

## Changes
- Fixed inconsistent bookmark data format (JSON strings → plain markdown)
- Extracted advocacy/skepticism rendering functions to separate utility file
- Fixed TypeScript compilation errors in test files
- Made advocacy/skepticism run unconditionally (API breakage fix)
- Improved code organization and reusability

## Technical Details
- Moved renderAdvocacyContent and renderSkepticismContent to utils/contentRenderers.tsx
- Fixed missing properties in test mocks (score_delta, etc)
- Corrected import paths for types in test files
- Ensured backward compatibility for API consumers
